common/persistence/fdb/storage/history_node_test.go
package storage

import (
	"math/big"
	"reflect"
	"testing"

	"github.com/apple/foundationdb/bindings/go/src/fdb"
	"github.com/apple/foundationdb/bindings/go/src/fdb/directory"
)

func TestHistoryNodeSaveAndLoad(t *testing.T) {
	fdb.MustAPIVersion(710)
	db := fdb.MustOpenDefault()
	dir, err := directory.CreateOrOpen(db, []string{"test"}, nil)
	if err != nil {
		t.Fatal(err)
	}
	ss := dir.Sub("executions")
	defer func() {
		db.Transact(func(tx fdb.Transaction) (interface{}, error) {
			tx.ClearRange(ss)

			return nil, nil
		})
	}()

	expected := &HistoryNode{
		ShardID:      42,
		TreeID:       []byte("a-tree-id"),
		BranchID:     []byte("a-branch-id"),
		NodeID:       big.NewInt(51),
		TxnID:        big.NewInt(18),
		PrevTxnID:    big.NewInt(21),
		Data:         []byte("some-data"),
		DataEncoding: "some-data-encoding",
	}
	err = expected.Save(db, ss)
	if err != nil {
		t.Fatal(err)
	}

	repo := NewHistoryNodeRepository(db, ss)
	actual, err := repo.Find(42, []byte("a-tree-id"), []byte("a-branch-id"), big.NewInt(51), big.NewInt(18))
	if err != nil {
		t.Fatal(err)
	}

	if !reflect.DeepEqual(expected, actual) {
		t.Errorf("didn't load expected serialized data from the database\n\texpected: %#v\n\tgot:      %#v", expected, actual)
	}
}
